test(sandbox): OCR round-2 fixes and #2114-compliant launch-release split for #3542
Final remediation batch for the nightly-failure effort:

- Apply the six OCR round-2 findings: sorted src-set equality for the venv
  init mounts, deduped ln -sfn notes in checkpoint storage, unconditional
  .git/info/attributes write in the persistence integration fixture
  (matching production), expect.objectContaining for the network exec
  assertions so added exec options keep the M1 env contract, and removal
  of the dead win32 branch in assertSignalDeath.
- Replace the rejected per-file eslint max-lines override (forbidden for
  packages/cli by the #2114 policy guard) with a real fix: move the pure
  launch-release helpers, types, and the podman connection constant into
  sandbox-launch-release.test-helpers.ts, following the existing
  sandbox-seatbelt.test-helpers.ts convention. The cast-based
  completedEngineProcess mock stays in the test file (production-scoped
  helper module may not use escape-hatch casts). Test file is back under
  the 800 effective-line cap with the standard gate.
- Record dispatch-4 nightly evidence and the TestRig [rest] timeout
  classification in the plan document.

Verification: cycle 12 format/lint/typecheck/build/smoke/diffcheck green;
full test suite green on re-run (cycle 12b; first run hit a proven-flaky
subagent termination test that passes 12/12 standalone and is untouched by
this PR); launch-release suite 12/12; eslint policy guard passes.
